[mod] last

+++ 00_Libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: chhan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/23 11:14:43 by chhan             #+#    #+#             */
-/*   Updated: 2020/12/30 15:45:05 by chan             ###   ########.fr       */
+/*   Updated: 2021/01/06 16:56:54 by chan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_memcmp(const void *ptr1, const void *ptr2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
-	while (++i < (int)n)
+	while (++i < n)
 	{
 		if (((unsigned char *)ptr1)[i] != ((unsigned char *)ptr2)[i])
 			return (((unsigned char *)ptr1)[i] - ((unsigned char *)ptr2)[i]);
